Implement fallback frame format conversion for spritesheet generation
