+++ src/lib/spritesheets/ktx2FileCache.js
@@ -156,5 +156,5 @@ export class Ktx2FileCache {
 }
 
 /**
- * @typedef {{sprites: import("./packFrames").SpriteData[], frameRate: number, scale: number}} CachedSpriteData
+ * @typedef {{sprites: import("./FramePacker").SpriteData[], frameRate: number, scale: number}} CachedSpriteData
  */
